kb.c: I made a stupid mistake, thanks Evandro, classify.c , just revert, too tired to fix and hate VC too much.
git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk/sphinx3@4214 c60273ff-4a12-0410-8d58-82ceea6d4170

+++ src/libs3decoder/libsearch/kb.c
@@ -334,8 +334,8 @@ void kb_init (kb_t *kb)
 				cmd_ln_int32("-dist_ds"),
 				cmd_ln_int32("-gs4gs"),
 				cmd_ln_int32("-svq4svq"),
-				(float32)cmd_ln_float64("-subvqbeam"),
-				(float32)cmd_ln_float64("-ci_pbeam"),
+				cmd_ln_float64("-subvqbeam"),
+				cmd_ln_float64("-ci_pbeam"),
 				cmd_ln_int32("-maxcdsenpf"),
 				kb->kbcore->mdef->n_ci_sen
 				);
